requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiosqlite>=0.20.0
aiohttp>=3.9.0
brotli>=1.1.0
fastapi>=0.110.0
uvicorn>=0.27.0